Split KEYCODES from KeyActionMapper
Move KEYCODES and related functions to a separate file. No further
refactoring is done yet.
