+++ src/js/game/hud/parts/keybinding_overlay.js
@@ -1,12 +1,7 @@
+import { getStringForKeyCode, KEYCODE_LMB, KEYCODE_MMB, KEYCODE_RMB } from "@/core/keycodes";
 import { makeDiv } from "../../../core/utils";
 import { T } from "../../../translations";
-import {
-    getStringForKeyCode,
-    KEYCODE_LMB,
-    KEYCODE_MMB,
-    KEYCODE_RMB,
-    KEYMAPPINGS,
-} from "../../key_action_mapper";
+import { KEYMAPPINGS } from "../../key_action_mapper";
 import { enumHubGoalRewards } from "../../tutorial_goals";
 import { BaseHUDPart } from "../base_hud_part";
 import { DynamicDomAttach } from "../dynamic_dom_attach";
